git: only link diff-highlight if not already linked

+++ git/install.osx.sh
@@ -6,4 +6,4 @@
 
 (cd $DOTFILES_ROOT/git/stow && stow -t $HOME/.gnupg .gnupg)
 
-ln -s $HOMEBREW_ROOT/opt/git/share/git-core/contrib/diff-highlight/diff-highlight $HOMEBREW_ROOT/bin/
+[ ! -e $HOMEBREW_ROOT/bin/diff-highlight ] && ln -s $HOMEBREW_ROOT/opt/git/share/git-core/contrib/diff-highlight/diff-highlight $HOMEBREW_ROOT/bin/
